ForbidDynamicProperties: fix the bugs
This commit updates the proof of concept trait to a fully working trait.

The tests on PHP 7.2 - 8.2 now pass, but more work is still needed to make the code compatible with PHP < 7.2.

This is basically a full rewrite of the trait based on the ideas contained in the proof of concept.

It solves the following problems:
* Overreach of the original `propertyExists()` method, which reported `true` for `private` properties everywhere in the class hierarchy, while for `private` properties not accessible from the calling context, the `OutOfBounds` exception should be thrown.
* Failure to set `private` properties on classes in the current class hierarchy, but not in the class containing the `trait`. Those will now be set using `Reflection`.
* Overreach when trying to set properties. Properties which PHP would report as inaccessible, should still be reported as such. To that effect, a `trigger_error()` call emulating the PHP native error has been added.

Note: While there are still some PHP > 7.2 hickups to smooth out, this should get us a lot closer to the intended behaviour.

Includes:
* Adding documentation to the trait in all the appropriate places.
* Adding a prefix to all properties in the `trait` to prevent potential conflicts wth properties declared in the class which composes-in the trait.
* Adding a prefix to all methods in the `trait`, with the exception of `__set()` (of course). This should minimize the risk of a class which uses the trait accidentally overloading one of the methods needed for the trait.